ajustes do jinq criaçao de implements

+++ agenda_api/src/main/java/br/gdtech/agenda/controller/AgendamentoController.java
@@ -21,7 +21,7 @@ public class AgendamentoController {
 	private AgendamentoService service;
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-	public Optional<Agendamento> findById(@PathParam("id") Integer id) {
+	public Optional<Agendamento> findById(@PathParam("id") Long id) {
 		return service.findById(id);
 	}
