+++ agenda_api/src/main/java/br/gdtech/agenda/controller/ProdutoServicoController.java
@@ -21,7 +21,7 @@ public class ProdutoServicoController {
 	private ProdutoServicoService service;
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-	public Optional<ProdutoServico> findById(@PathParam("id") Integer id) {
+	public Optional<ProdutoServico> findById(@PathParam("id") Long id) {
 		return service.findById(id);
 	}
